Add files via upload
Tic-Tac-Toe is a two player game in which each player marks the cell with 0 and 1 with 9 in the 9 cells of a 3*3 board alternatively. The winner is the one who succeeds to mark their pattern either in a row or column or diagonal. It can be used to represent game playing technique in AI.